switch to u8 return, remove cast

+++ include/d/com/d_com_inf_game.h
@@ -1005,7 +1005,7 @@ inline u8 dComIfGs_checkBottle(u8 i_itemNo) {
     return g_dComIfG_gameInfo.info.getPlayer().getItem().checkBottle(i_itemNo);
 }
 
-inline u32 dComIfGs_checkOptVibration() {
+inline u8 dComIfGs_checkOptVibration() {
     return g_dComIfG_gameInfo.info.getPlayer().getConfig().checkVibration();
 }
